requirements.txt
streamlit==1.34.0
langchain==0.2.0
langchain-community==0.2.0
langchain-groq==0.1.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.1